Refactor middleware for token verification and access control (v0.1.14)
In this commit, I updated the middleware to enhance token verification and access control. The main changes include replacing the token verification method with jwtVerify for more reliable user authentication, updating the logic for handling admin pages, including the handling of expired tokens, and adding error handling during token verification, which improves the application's security. Additionally, a temporary component has been added to the student page to inform users that this section will be available soon; this is a temporary solution, and the page will be further developed and ready for use in the future. Furthermore, this commit includes the addition of verification tokens for Google and Yandex, improving integration with analytics tools. This commit is part of an update for production and will be pushed to the appropriate branch.